[Chore] 추후 실습을 위해 CounterApp 컴포넌트의 실습 파일을 components 하위 counterApp 경로로 이동
