+++ src/components/counterApp/atoms/Count/index.stories.tsx
@@ -1,21 +1,21 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Title from '.'
+import Count from '.'
 
 const meta = {
-  title: 'Atoms/Title',
-  component: Title,
+  title: 'CounterApp/Atoms/Count',
+  component: Count,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof Title>
+} satisfies Meta<typeof Count>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
   args: {
-    title: '카운터 앱',
+    value: 0,
   },
 }
