+++ src/components/counterApp/organisms/Counter/index.stories.tsx
@@ -1,21 +1,17 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Button from '.'
+import Counter from '.'
 
 const meta = {
-  title: 'Atoms/Button',
-  component: Button,
+  title: 'CounterApp/Organisms/Counter',
+  component: Counter,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof Button>
+} satisfies Meta<typeof Counter>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
-  args: {
-    label: '버튼',
-  },
-}
+export const Default: Story = {}
