+++ src/components/counterApp/pages/Home/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import CounterApp from 'components/templates/CounterApp'
+import CounterApp from '@/components/counterApp/templates/CounterApp'
 
 const Home = () => <CounterApp />
