+++ src/components/counterApp/templates/CounterApp/index.stories.tsx
@@ -1,21 +1,17 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Count from '.'
+import CounterApp from '.'
 
 const meta = {
-  title: 'Atoms/Count',
-  component: Count,
+  title: 'CounterApp/Organisms/CounterApp',
+  component: CounterApp,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof Count>
+} satisfies Meta<typeof CounterApp>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
-  args: {
-    value: 0,
-  },
-}
+export const Default: Story = {}
