+++ src/components/counterApp/templates/CounterApp/index.tsx
@@ -1,5 +1,5 @@
-import Title from 'components/atoms/Title'
-import Counter from 'components/organisms/Counter'
+import Title from '@/components/counterApp/atoms/Title'
+import Counter from '@/components/counterApp/organisms/Counter'
 
 const CounterApp: React.FC = () => (
   <div className='flex flex-col items-center justify-center h-screen'>
